refactor: remove json feature
