+++ Cargo.toml
@@ -14,16 +14,15 @@ exclude = [
 ]
 
 [features]
-json = ["serde", "serde_json", "url/serde", "indexmap/serde"]
 logging = ["log"]
-default = ["json", "logging"]
+default = ["logging"]
 
 [dependencies]
-indexmap = { version = "1.7.0" }
+indexmap = { version = "1.7.0", features = ["serde"] }
 log = { version = "0.4.14", optional = true }
-serde = { version = "1.0.129", features = ["derive"], optional = true }
-serde_json = { version = "1.0.66", features = ["preserve_order"], optional = true }
-url = { version = "2.2.2" }
+serde = { version = "1.0.129", features = ["derive"] }
+serde_json = { version = "1.0.66", features = ["preserve_order"] }
+url = { version = "2.2.2", features = ["serde"] }
 
 [dev-dependencies]
 walkdir = "2.3.2"
